producer_app/requirements.txt
mcap
mcap-protobuf-support
pulsar-client
pyarrow
